UPDATE: Add validation so that it is not possible to delete an order id that already has a transaction.

+++ app/Http/Controllers/Service/OrderDetailController.php
@@ -359,6 +359,20 @@ class OrderDetailController extends Controller
                 );
             }
 
+            $transaction = Transaction::where('order_detail_id', $orderDetail->id)
+                ->first();
+            if ($transaction) {
+                return response()->json(
+                    new WithoutDataResource(
+                        Response::HTTP_BAD_REQUEST,
+                        'ORDER_HAS_TRANSACTION',
+                        'Tidak Dapat Dihapus',
+                        'Order detail ini sudah memiliki transaksi dan tidak dapat dihapus.'
+                    ),
+                    Response::HTTP_BAD_REQUEST
+                );
+            }
+
             $orderDetail->delete();
 
             DB::commit();
